requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic-settings>=2.0.0
pyyaml>=6.0
